+++ Firestore/Swift/CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased
+- [added] Added Ascync/Await wrapper for `CollectionReference.addDocument()`.
+
 # v8.9.0
 - [added] Added `@FirestoreQuery` property wrapper for querying data from a
   Firestore collection.